bitlpm: Move fuzzer to dedicated file
The fuzzer test workflows we have rely on the fuzzers being in a
filename that ends with 'fuzz_test.go'. Move the fuzzer there to make
sure that any changes to the fuzzer will trigger the test workflows.
